replace xcm filter and omniexchange to adapter
