+++ runtime/adapters/Cargo.toml
@@ -10,6 +10,7 @@ repository = "https://github.com/galacticcouncil/warehouse/tree/master/adapters"
 [dependencies]
 codec = { default-features = false, features = ["derive"], package = "parity-scale-codec", version = "3.4.0" }
 log = { version = "0.4.17", default-features = false }
+scale-info = { version = "2.3.1", default-features = false, features = ["derive"] }
 
 # HydraDX dependencies
 primitives = { workspace = true }
@@ -22,6 +23,8 @@ pallet-circuit-breaker = { workspace = true }
 warehouse-liquidity-mining = { workspace = true }
 pallet-omnipool-liquidity-mining = { workspace = true }
 pallet-dynamic-fees = { workspace = true }
+pallet-route-executor = { workspace = true }
+pallet-currencies = { workspace = true }
 
 # Substrate dependencies
 frame-support = { workspace = true }
@@ -29,6 +32,8 @@ frame-system = { workspace = true }
 sp-runtime = { workspace = true }
 sp-std = { workspace = true }
 primitive-types = { workspace = true }
+sp-core = { workspace = true }
+sp-io = { workspace = true }
 
 # Polkadot dependencies
 polkadot-xcm = { workspace = true }
@@ -38,9 +43,19 @@ xcm-executor = { workspace = true }
 # ORML dependencies
 orml-xcm-support = { workspace = true }
 orml-traits = { workspace = true }
+orml-utilities = { workspace = true }
+orml-tokens = { workspace = true }
+
+# Pallets
+pallet-balances = { workspace = true }
+
+# Cumulus dependencies
+cumulus-primitives-core = { workspace = true }
+
 
 [dev-dependencies]
 lazy_static = { features = ["spin_no_std"], version = "1.4.0" }
+pretty_assertions = "1.4.0"
 
 [features]
 default = ["std"]
