+++ runtime/adapters/src/tests/mod.rs
@@ -1,2 +1,4 @@
 pub mod mock;
+pub mod trader;
 pub mod xcm_exchange;
+pub mod xcm_filter;
